requirements.txt
fastapi==0.117.1
uvicorn==0.37.0
pydantic==2.11.9
pydantic-core==2.33.2
pydantic-settings==2.11.0
python-dotenv==1.1.1
sqlalchemy==2.0.43
sqlmodel==0.0.22
httpx==0.27.2
pytest==8.4.2
pytest-asyncio==0.26.0
pyodbc==5.2.0
cryptography==46.0.1
pyjwt==2.10.1
starlette==0.48.0
typing-extensions==4.15.0
fastapi-azure-auth==5.2.0
requests==2.32.5
